Reformatted & changed current calculation
Code was reformatted to be dynamic with real-time changes.  Calculations were changed to sample a period slightly longer than on 50hz cycle (for EU, US is 60hz)  to compute ADC value.  This is then repeated 5 times with the highest and lowest being removed before being averaged.  Overall calculation time is now approximately 102.25ms.  A cycle time was then added to limit Wi-Fi activity and preserve server bandwidth independently of the time spent reading samples.  Lastly LED pins were updated and animation was made dynamic with the rest of the code.